requirements.txt
django>=4.0
djangorestframework
psycopg2-binary
python-decouple
pandas
openpyxl
django-cors-headers
drf-yasg

